Add files via upload
This is a shiny app to run the AI4HealtyAging risk calculator. The calculator will offer you the probabilities of Prostate cancer (PCa), and results will be proposed as favorable (green) or unfavorable (red) after applying the Threshold probability selected. Using the threshold point, the PCa wrongly classified and biopsied avoided below the cutoff point are displayed in clinical utility curve for all data.